move load data to datagen
